# .docs/api/swagger.yaml
definitions:
  data.LoginUserAccountParam:
    properties:
      password:
        type: string
      username_or_email:
        type: string
    required:
    - password
    - username_or_email
    type: object
  data.LoginUserAccountResult:
    properties:
      token:
        type: string
      username:
        type: string
    type: object
  data.RegisterUserAccountParam:
    properties:
      email:
        type: string
      family_name:
        type: string
      given_name:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
      username:
        type: string
    type: object
  data.RegisterUserAccountResult:
    properties:
      email:
        type: string
      family_name:
        type: string
      given_name:
        type: string
      username:
        type: string
    type: object
  data.ShowUserAccountResult:
    properties:
      email:
        type: string
      family_name:
        type: string
      given_name:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: WebApp API Doc
  version: "1.0"
paths:
  /user_account:
    get:
      consumes:
      - application/json
      description: Show a logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.ShowUserAccountResult'
        "401":
          description: ""
      summary: Show a logged in user
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: RegisterUserAccountParam
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/data.RegisterUserAccountParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.RegisterUserAccountResult'
        "400":
          description: ""
      summary: Register a user
  /user_account/login:
    post:
      consumes:
      - application/json
      description: login with user
      parameters:
      - description: LoginUserAccountParam
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/data.LoginUserAccountParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.LoginUserAccountResult'
        "400":
          description: ""
        "401":
          description: ""
      summary: Login with user
swagger: "2.0"